+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [1.176.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.176.0...v1.176.1) (2025-08-19)
+
+
+### Bug Fixes
+
+* **drawer:** L3-7407 explicitly unset pointer-events ([#691](https://github.com/PhillipsAuctionHouse/seldon/issues/691)) ([233d8e6](https://github.com/PhillipsAuctionHouse/seldon/commit/233d8e6c6c0cbe9268a99cbe5d6791ae3bd0b375))
+
+# [1.176.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.175.0...v1.176.0) (2025-08-18)
+
+
+### Features
+
+* **countryPicker:** L3-7243 Build country and country code picker variants ([#677](https://github.com/PhillipsAuctionHouse/seldon/issues/677)) ([19e540e](https://github.com/PhillipsAuctionHouse/seldon/commit/19e540edbf0037b00094de00231a4e2d7ec1c66e))
+
+# [1.175.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.174.0...v1.175.0) (2025-08-15)
+
+
+### Features
+
+* **stepper:** L3-7123 Create progress component (Part 1) ([#668](https://github.com/PhillipsAuctionHouse/seldon/issues/668)) ([ecce47a](https://github.com/PhillipsAuctionHouse/seldon/commit/ecce47a63fbe139969adeba8aa17dd0a5623984b))
+
+# [1.174.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.173.0...v1.174.0) (2025-08-14)
+
+
+### Features
+
+* **detailedRadio:** L3-7240 Make Radio input selector match figma for primary details (Bidder type select) ([#671](https://github.com/PhillipsAuctionHouse/seldon/issues/671)) ([ed88be9](https://github.com/PhillipsAuctionHouse/seldon/commit/ed88be98a08aafce98b43ae41f4386449e72c651))
+
 # [1.173.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.172.0...v1.173.0) (2025-08-12)